AI/Classification/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
openai==1.3.0
python-dotenv==1.0.0
uvicorn==0.24.0
httpx==0.25.1
